+++ packages/auth/__tests__/providers/cognito/updatePassword.test.ts
@@ -8,7 +8,8 @@ import { AuthError } from '../../../src/errors/AuthError';
 import { AuthValidationErrorCode } from '../../../src/errors/types/validation';
 import { updatePassword } from '../../../src/providers/cognito';
 import { ChangePasswordException } from '../../../src/providers/cognito/types/errors';
-import { changePassword } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
+import { createChangePasswordClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../../../src/providers/cognito/factories';
 
 import { getMockError, mockAccessToken } from './testUtils/data';
 import { setUpGetConfig } from './testUtils/setUpGetConfig';
@@ -22,15 +23,22 @@ jest.mock('@aws-amplify/core/internals/utils', () => ({
 	isBrowser: jest.fn(() => false),
 }));
 jest.mock(
-	'../../../src/providers/cognito/utils/clients/CognitoIdentityProvider',
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
 );
+jest.mock('../../../src/providers/cognito/factories');
 
 describe('updatePassword', () => {
 	const oldPassword = 'oldPassword';
 	const newPassword = 'newPassword';
 	// assert mocks
 	const mockFetchAuthSession = fetchAuthSession as jest.Mock;
-	const mockChangePassword = changePassword as jest.Mock;
+	const mockChangePassword = jest.fn();
+	const mockCreateChangePasswordClient = jest.mocked(
+		createChangePasswordClient,
+	);
+	const mockCreateCognitoUserPoolEndpointResolver = jest.mocked(
+		createCognitoUserPoolEndpointResolver,
+	);
 
 	beforeAll(() => {
 		setUpGetConfig(Amplify);
@@ -41,11 +49,13 @@ describe('updatePassword', () => {
 
 	beforeEach(() => {
 		mockChangePassword.mockResolvedValue({});
+		mockCreateChangePasswordClient.mockReturnValueOnce(mockChangePassword);
 	});
 
 	afterEach(() => {
 		mockChangePassword.mockReset();
 		mockFetchAuthSession.mockClear();
+		mockCreateChangePasswordClient.mockClear();
 	});
 
 	it('should call changePassword', async () => {
@@ -59,6 +69,25 @@ describe('updatePassword', () => {
 				ProposedPassword: newPassword,
 			}),
 		);
+	});
+
+	it('invokes mockCreateCognitoUserPoolEndpointResolver with expected endpointOverride', async () => {
+		const expectedUserPoolEndpoint = 'https://my-custom-endpoint.com';
+		jest.mocked(Amplify.getConfig).mockReturnValueOnce({
+			Auth: {
+				Cognito: {
+					userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
+					userPoolId: 'us-west-2_zzzzz',
+					identityPoolId: 'us-west-2:xxxxxx',
+					userPoolEndpoint: expectedUserPoolEndpoint,
+				},
+			},
+		});
+		await updatePassword({ oldPassword, newPassword });
+
+		expect(mockCreateCognitoUserPoolEndpointResolver).toHaveBeenCalledWith({
+			endpointOverride: expectedUserPoolEndpoint,
+		});
 	});
 
 	it('should throw an error when oldPassword is empty', async () => {