+++ packages/storage/src/providers/s3/types/index.ts
@@ -2,23 +2,21 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export {
-	GetUrlOptionsWithKey,
-	GetUrlOptionsWithPath,
-	UploadDataOptionsWithPath,
-	UploadDataOptionsWithKey,
-	GetPropertiesOptionsWithKey,
-	GetPropertiesOptionsWithPath,
-	ListAllOptionsWithPrefix,
-	ListPaginateOptionsWithPrefix,
-	ListAllOptionsWithPath,
-	ListPaginateOptionsWithPath,
+	GetUrlWithKeyOptions,
+	GetUrlWithPathOptions,
+	UploadDataWithPathOptions,
+	UploadDataWithKeyOptions,
+	GetPropertiesWithKeyOptions,
+	GetPropertiesWithPathOptions,
+	ListAllWithPrefixOptions,
+	ListPaginateWithPrefixOptions,
+	ListAllWithPathOptions,
+	ListPaginateWithPathOptions,
 	RemoveOptions,
-	DownloadDataOptionsWithPath,
-	DownloadDataOptionsWithKey,
-	CopyDestinationOptionsWithKey,
-	CopySourceOptionsWithKey,
-	CopyWithPathSourceOptions,
-	CopyWithPathDestinationOptions,
+	DownloadDataWithPathOptions,
+	DownloadDataWithKeyOptions,
+	CopyDestinationWithKeyOptions,
+	CopySourceWithKeyOptions,
 } from './options';
 export {
 	UploadDataOutput,
@@ -58,4 +56,3 @@ export {
 	ListAllWithPathInput,
 	ListPaginateWithPathInput,
 } from './inputs';
-export { S3Exception } from './errors';